requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
openai==1.56.1
requests==2.32.3
Faker==33.1.0